@pellared 2024-03-15
> @pellared I am curious how does this compare with "sort/dedup" approach that can be done in-place and requires no maps and no allocations. Did you by any chance try that?

`log.Record` and `Record` are different structs so we need to copy the data anyway. This is the code that converts API representation  LogRecord  (`log.Record`) to SDK representation of LogRecord (`Record`).

Also sorting itself would make it `O(n*log(n))`. This is looks `O(n)`.